+++ core/src/main/java/nl/andrewlalis/aos_core/model/Team.java
@@ -8,8 +8,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Team implements Serializable {
-	public static final double SPAWN_RADIUS = 3;
-	public static final double SUPPLY_POINT_RADIUS = 2;
+	public static final float SPAWN_RADIUS = 3;
+	public static final float SUPPLY_POINT_RADIUS = 2;
 
 	private final String name;
 	private final java.awt.Color color;
